# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Träwelling API",
        "description": "Träwelling user API description. This is an incomplete documentation with still many errors. The API is currently not yet stable. Endpoints are still being restructured. Both the URL and the request or body can be changed. Breaking changes will be announced on GitHub: https://github.com/Traewelling/traewelling/blob/develop/API_CHANGELOG.md",
        "contact": {
            "email": "support@traewelling.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0 - alpha"
    },
    "servers": [
        {
            "url": "https://traewelling.de/api/v1",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "This instance"
        }
    ],
    "paths": {
...
        "/settings/profile": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Get the current user's profile settings",
                "description": "Get the current user's profile settings",
                "operationId": "getProfileSettings",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserProfileSettingsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "read-settings"
                        ]
                    },
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Settings"
                ],
                "summary": "Update the current user's profile settings",
                "description": "Update the current user's profile settings",
                "operationId": "updateProfileSettings",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "maxLength": 25,
                                        "example": "gertrud123"
                                    },
                                    "displayName": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "Gertrud"
                                    },
                                    "privateProfile": {
                                        "type": "boolean",
                                        "example": false,
                                        "nullable": true
                                    },
                                    "preventIndex": {
                                        "type": "boolean",
                                        "example": false,
                                        "nullable": true
                                    },
                                    "privacyHideDays": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "defaultStatusVisibility": {
                                        "oneOf": [
                                            {
                                                "$ref": "#/components/schemas/StatusVisibility"
                                            }
                                        ],
                                        "nullable": true
                                    },
                                    "mastodonVisibility": {
                                        "oneOf": [
                                            {
                                                "$ref": "#/components/schemas/MastodonVisibility"
                                            }
                                        ],
                                        "nullable": true
                                    },
                                    "mapProvider": {
                                        "oneOf": [
                                            {
                                                "$ref": "#/components/schemas/MapProvider"
                                            }
                                        ],
                                        "nullable": true
                                    },
                                    "friendCheckin": {
                                        "oneOf": [
                                            {
                                                "$ref": "#/components/schemas/FriendCheckinSetting"
                                            }
                                        ],
                                        "nullable": true
                                    },
                                    "likesEnabled": {
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    },
                                    "pointsEnabled": {
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserProfileSettingsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "passport": [
                            "write-settings"
                        ]
                    },
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MapProvider": {
                "title": "MapProvider",
                "description": "What type of map provider (cargo, open-railway-map) did the user specify?",
                "type": "string",
                "enum": [
                    "cargo",
                    "open-railway-map"
                ],
                "example": "cargo"
            },
            "MastodonVisibility": {
                "title": "visibility",
                "description": "What type of visibility (0=public, 1=unlisted, 2=followers, 3=private) did the user specify for\n *     future posts to Mastodon? Some instances such as chaos.social discourage bot posts on public timelines.",
                "type": "integer",
                "enum": [
                    0,
                    1,
                    2,
                    3
                ],
                "example": 1
            },
...
            "StatusVisibility": {
                "title": "visibility",
                "description": "What type of visibility (0=public, 1=unlisted, 2=followers, 3=private, 4=authenticated) did the\n *      user specify?",
                "type": "integer",
                "enum": [
                    0,
                    1,
                    2,
                    3,
                    4
                ],
                "example": 0
            },
...
            "FriendCheckinSetting": {
                "title": "FriendCheckinSetting",
                "type": "string",
                "enum": [
                    "forbidden",
                    "friends",
                    "list"
                ],
                "example": "forbidden"
            },
...
            "UserProfileSettingsResource": {
                "title": "UserProfileSettings",
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "Gertrud123"
                    },
                    "displayName": {
                        "type": "string",
                        "example": "Gertrud"
                    },
                    "profilePicture": {
                        "type": "string",
                        "example": "https://traewelling.de/@Gertrud123/picture"
                    },
                    "privateProfile": {
                        "type": "boolean",
                        "example": false
                    },
                    "preventIndex": {
                        "description": "Did the user choose to prevent search engines from indexing their profile?",
                        "type": "boolean",
                        "example": false
                    },
                    "defaultStatusVisibility": {
                        "$ref": "#/components/schemas/StatusVisibility"
                    },
                    "privacyHideDays": {
                        "description": "Number of days to hide the user's location history",
                        "type": "integer",
                        "example": 1
                    },
                    "password": {
                        "type": "boolean",
                        "example": true
                    },
                    "email": {
                        "type": "string",
                        "example": "gertrud@traewelling.de"
                    },
                    "emailVerified": {
                        "type": "boolean",
                        "example": true
                    },
                    "profilePictureSet": {
                        "type": "boolean",
                        "example": true
                    },
                    "mastodon": {
                        "type": "string",
                        "example": "https://mastodon.social/@Gertrud123"
                    },
                    "mastodonVisibility": {
                        "$ref": "#/components/schemas/MastodonVisibility"
                    },
                    "friendCheckin": {
                        "$ref": "#/components/schemas/FriendCheckinSetting"
                    },
                    "likesEnabled": {
                        "type": "boolean",
                        "example": true
                    },
                    "pointsEnabled": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "OAuth2 authorizationCode Flow. Get your token from https://traewelling.de/settings/applications. Set the redirect URL to https://traewelling.de/api/oauth2-callback to use it here. See also https://laravel.com/docs/9.x/passport",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/auth/refresh",
                        "scopes": {
                            "read-statuses": "see all statuses",
                            "read-notifications": "see your notifications",
                            "read-statistics": "see your statistics",
                            "read-search": "search in Träwelling",
                            "write-statuses": "create, edit, delete statuses",
                            "write-likes": "create and remove likes",
                            "write-notifications": "mark notifications as read, clear notifications",
                            "write-exports": "request data exports",
                            "write-follows": "follow and unfollow users",
                            "write-followers": "accept follow requests and remove followers",
                            "write-blocks": "block and unblock users, mute and unmute users",
                            "write-event-suggestions": "suggest events in your name",
                            "write-support-tickets": "create support tickets in your name",
                            "read-settings": "see your settings, email, etc.",
                            "write-settings-profile": "edit your profile",
                            "read-settings-profile": "see your profile data, e.g. email",
                            "write-settings-mail": "edit your email",
                            "write-settings-profile-picture": "edit your profile picture",
                            "write-settings-privacy": "change your privacy settings",
                            "read-settings-followers": "show follow-requests and followers",
                            "write-settings-calendar": "create and delete new calendar-tokens",
                            "extra-write-password": "change your password",
                            "extra-terminate-sessions": "log you out of other sessions and apps",
                            "extra-delete": "delete your account"
                        }
                    }
                }
            },
            "token": {
...
    },
...
}